refact: 스크립트 분리
